Improve color(at:) by using dispatch
Getting a color from an image happens in the main thread, this can
cause slowdowns if the method is used frequently (e.g. a feed). To
offload the main thread from extracting the pixel color of a specific
point, the method will now use a completion closure that will be
dispatched into the main thread when the color has been extracted.
Internally `color(at:)` runs in a user interactive queue.